refactor(core): replace queue/map/spell const ints with typed enums
Replace the LolQueueIds / LolMapIds / SummonerSpellIds static const-int
classes with LolQueueId / LolMapId / SummonerSpellId enums backed by the
same Riot numeric values. MainAnalysisOptions.QueueId becomes LolQueueId —
the real type-safety win on the config knob (it still binds from the int
"420" in appsettings).

The entity/DB layer stays int (Match.QueueId/MapId, Summoner*Id are int
columns), so the enums cross that boundary with an explicit (int) cast at
each call site (~20 spots across query services, ingestor processes,
SummonerSpellPair, and the test builders/fixtures). No deprecated shim:
TreatWarningsAsErrors makes [Obsolete] a build break, so this is the
all-at-once migration agreed in review.

239 unit + 77 integration tests green.

Closes #280

+++ backend/Ingestor/Processes/ChampionPatternAggregationProcess.cs
@@ -15,7 +15,7 @@ public sealed class ChampionPatternAggregationProcess(
 
     public async Task<object?> RunCoreAsync(CancellationToken ct)
     {
-        var queueId = analysisOptions.Value.QueueId;
+        var queueId = (int)analysisOptions.Value.QueueId;
         var aggregationInputs = await sourceRowReader.LoadAggregationInputsAsync(queueId, ct);
         if (aggregationInputs.SourceRows.Count == 0 && aggregationInputs.ExistingAggregateScopes.Count == 0)
         {
